Moved `VeniceReentrantReadWriteLock` from `venice.service.venice-controller` into `internal/venice-client-common` and used it in `PartitionConsumptionState`. 🍿
